pyproject.toml
[project]
classifiers = [
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Software Development :: Quality Assurance",
  "Topic :: Software Development :: Testing",
]
description = "Patternfly4 widget library for Widgetastic."
dynamic = ["version"]
keywords = ["widgetastic", "patternfly"]
license = {file = "LICENSE"}
maintainers = [{name = "Dmitrii Misharov", email = "misharov@redhat.com"}]
name = "widgetastic.patternfly4"
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
  "widgetastic.core>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pre-commit",
  "pytest",
  "pytest-cov",
  "pytest-xdist",
  "codecov",
]
doc = ["sphinx"]

[project.urls]
repository = "https://github.com/RedHatQE/widgetastic.patternfly4"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling", "hatch-vcs"]

[tool.hatch.build.targets.wheel]
packages = ["src/widgetastic_patternfly4"]

[tool.hatch.version]
source = "vcs"
